requirements.txt
transformers
langchain
langchain-huggingface
python-dotenv
torch